~ WIP new browse API and implementation
